@@ -14,42 +14,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.6",
+    "@angular/animations": "^10.1.4",
     "@angular/cdk": "^9.1.6",
-    "@angular/common": "^9.1.6",
-    "@angular/compiler": "^9.1.6",
-    "@angular/core": "^9.1.6",
-    "@angular/forms": "^9.1.6",
+    "@angular/common": "^10.1.4",
+    "@angular/compiler": "^10.1.4",
+    "@angular/core": "^10.1.4",
+    "@angular/forms": "^10.1.4",
     "@angular/material": "^9.2.3",
-    "@angular/platform-browser": "^9.1.6",
-    "@angular/platform-browser-dynamic": "^9.1.6",
-    "@angular/router": "^9.1.6",
+    "@angular/platform-browser": "^10.1.4",
+    "@angular/platform-browser-dynamic": "^10.1.4",
+    "@angular/router": "^10.1.4",
     "core-js": "^3.6.5",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.5.3",
-    "zone.js": "^0.10.2"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.5",
-    "@angular/cli": "^9.1.5",
-    "@angular/compiler-cli": "^9.1.6",
+    "@angular-devkit/build-angular": "^0.1001.4",
+    "@angular/cli": "^10.1.4",
+    "@angular/compiler-cli": "^10.1.4",
     "@types/jasmine": "~3.5.10",
     "@types/node": "~13.13.5",
     "codelyzer": "~5.2.2",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "^5.0.5",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^3.0.2",
-    "karma-jasmine": "~3.1.1",
-    "karma-jasmine-html-reporter": "^1.5.3",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.16",
     "prettier": "^2.0.5",
-    "protractor": "^6.0.0",
+    "protractor": "~7.0.0",
     "ts-node": "~8.10.1",
-    "tslint": "~6.1.2",
-    "typescript": "~3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.3"
   }
 }
